Move SettingsExternalURLs to cardstack/constants
